+++ src/main/java/br/com/zalf/prolog/webservice/Injection.java
@@ -4,6 +4,8 @@ import br.com.zalf.prolog.webservice.colaborador.ColaboradorDao;
 import br.com.zalf.prolog.webservice.colaborador.ColaboradorDaoImpl;
 import br.com.zalf.prolog.webservice.frota.checklist.ChecklistDao;
 import br.com.zalf.prolog.webservice.frota.checklist.ChecklistDaoImpl;
+import br.com.zalf.prolog.webservice.frota.checklist.ordemServico.OrdemServicoDao;
+import br.com.zalf.prolog.webservice.frota.checklist.ordemServico.OrdemServicoDaoImpl;
 import br.com.zalf.prolog.webservice.frota.pneu.afericao.AfericaoDao;
 import br.com.zalf.prolog.webservice.frota.pneu.afericao.AfericaoDaoImpl;
 import br.com.zalf.prolog.webservice.frota.pneu.pneu.PneuDao;
@@ -44,5 +46,9 @@ public final class Injection {
 
     public static PneuDao providePneuDao() {
         return new PneuDaoImpl();
+    }
+
+    public static OrdemServicoDao provideOrdemServicoDao() {
+        return new OrdemServicoDaoImpl();
     }
 }