+++ src/main/java/br/com/zalf/prolog/webservice/frota/checklist/ordemServico/OrdemServicoDao.java
@@ -4,6 +4,7 @@ import br.com.zalf.prolog.webservice.frota.checklist.model.Checklist;
 
 import java.sql.Connection;
 import java.sql.SQLException;
+import java.util.Date;
 import java.util.List;
 
 /**
@@ -28,6 +29,9 @@ public interface OrdemServicoDao {
 	//TODO - adicionar javadoc
 	List<ItemOrdemServico> getItensOsManutencaoHolder(String placa, String status,
 													  int limit, long offset, String prioridade) throws SQLException;
+
+	List<ItemOrdemServico> getItensOsManutencaoHolder(String placa, Date dataInicial, Date dataFinal,
+													  boolean itensCriticosRetroativos) throws SQLException;
 
 	/**
 	 * insere um item com problema na OS